Bury couple of static methods on triangle links
