+++ src/map-director/MapRegionChangesTask.cpp
@@ -223,9 +223,6 @@ void link_triangles
 
                 if (this_tri == other_tri) continue;
                 TriangleLink::attach_matching_points(this_tri, other_tri);
-#               if 0
-                this_tri->attempt_attachment_to(other_tri);
-#               endif
         }}
     }}
 }
